Fix bool, u8 and i8 conversions
I believe u8 and i8 were not supported because bool was converted to u8, and that caused conflicts. However in 2018 the C repr of bool was specified to be the same as _Bool in C, therefore we can just have bool convert to bool and then add proper support for u8 and i8.